Rename the BackupException to be BackupAndRestoreException
