+++ kin-recovery/recovery/src/main/java/kin/recovery/restore/presenter/RestoreEnterPasswordPresenterImpl.java
@@ -4,15 +4,15 @@ package kin.recovery.restore.presenter;
 import static kin.recovery.events.RestoreEventCode.RESTORE_PASSWORD_DONE_TAPPED;
 import static kin.recovery.events.RestoreEventCode.RESTORE_PASSWORD_ENTRY_PAGE_BACK_TAPPED;
 import static kin.recovery.events.RestoreEventCode.RESTORE_PASSWORD_ENTRY_PAGE_VIEWED;
-import static kin.recovery.exception.BackupException.CODE_RESTORE_FAILED;
-import static kin.recovery.exception.BackupException.CODE_RESTORE_INVALID_KEYSTORE_FORMAT;
+import static kin.recovery.exception.BackupAndRestoreException.CODE_RESTORE_FAILED;
+import static kin.recovery.exception.BackupAndRestoreException.CODE_RESTORE_INVALID_KEYSTORE_FORMAT;
 import static kin.recovery.restore.presenter.RestorePresenterImpl.KEY_ACCOUNT_KEY;
 
 import android.os.Bundle;
 import android.support.annotation.NonNull;
 import kin.recovery.Validator;
 import kin.recovery.events.CallbackManager;
-import kin.recovery.exception.BackupException;
+import kin.recovery.exception.BackupAndRestoreException;
 import kin.recovery.restore.view.RestoreEnterPasswordView;
 import kin.recovery.utils.Logger;
 import kin.sdk.KinAccount;
@@ -47,7 +47,7 @@ public class RestoreEnterPasswordPresenterImpl extends BaseChildPresenterImpl<Re
 		try {
 			KinAccount kinAccount = importAccount(keystoreData, password);
 			getParentPresenter().navigateToRestoreCompletedPage(kinAccount);
-		} catch (BackupException e) {
+		} catch (BackupAndRestoreException e) {
 			Logger.e("RestoreEnterPasswordPresenterImpl - restore failed.", e);
 			if (e.getCode() == CODE_RESTORE_INVALID_KEYSTORE_FORMAT) {
 				getView().invalidQrError();
@@ -58,18 +58,19 @@ public class RestoreEnterPasswordPresenterImpl extends BaseChildPresenterImpl<Re
 	}
 
 	private KinAccount importAccount(@NonNull final String keystore, @NonNull final String password)
-		throws BackupException {
+		throws BackupAndRestoreException {
 		Validator.checkNotNull(keystore, "keystore");
 		Validator.checkNotNull(keystore, "password");
 		KinAccount importedAccount;
 		try {
 			importedAccount = getParentPresenter().getKinClient().importAccount(keystore, password);
 		} catch (CryptoException e) {
-			throw new BackupException(CODE_RESTORE_FAILED, "Could not import the account");
+			throw new BackupAndRestoreException(CODE_RESTORE_FAILED, "Could not import the account");
 		} catch (CreateAccountException e) {
-			throw new BackupException(CODE_RESTORE_FAILED, "Could not create the account");
+			throw new BackupAndRestoreException(CODE_RESTORE_FAILED, "Could not create the account");
 		} catch (CorruptedDataException e) {
-			throw new BackupException(CODE_RESTORE_INVALID_KEYSTORE_FORMAT, "The keystore is invalid - wrong format");
+			throw new BackupAndRestoreException(CODE_RESTORE_INVALID_KEYSTORE_FORMAT,
+				"The keystore is invalid - wrong format");
 		}
 		return importedAccount;
 	}
